Continuous Authentication Feature addition
Added the entire mechanism for the CA on the project. In more detail it relies on the same mechanism as the user traning period used for the machine learning classier model data, however, there is not timer. Simply the CA continues to indefinitely to track the user behaviour on the dashboard and send the data to the ML classier by using the FLASK API.
Responses are sent back and upon such responses the user is logged out in case of a 0 response or continues logged in for 1 value responses.